+++ CHANGELOG.md
@@ -2,10 +2,11 @@
 
 ## 3.5.2
 ### Changed
-- Updated Kotlin to 1.8.10
+- Updated Kotlin to 1.8.22
 - 
 ### Fixed
 - Fix `CliktCommand.prompt` on NodeJS targets that would hang due to KT-55817 ([#387](https://github.com/ajalt/clikt/issues/387))
+- When `treatUnknownOptionsAsArgs` is true, grouped unknown short options will now be treated as arguments rather than reporting an error.
 
 ## 3.5.1
 ### Changed